+++ src/corpus/config/config.composable.ts
@@ -1,7 +1,6 @@
 import { computed } from "vue";
-import { useI18n } from "vue-i18n";
+import { computedAsync } from "@vueuse/core";
 import {
-  emptyConfig,
   makeConfig,
   parseConfig,
   type ConfigOptions,
@@ -9,31 +8,20 @@ import {
 import useLocale from "@/i18n/locale.composable";
 import useMinkBackend from "@/api/backend.composable";
 import { useResourceStore } from "@/store/resource.store";
-import useMessenger from "@/message/messenger.composable";
 
 export default function useConfig(corpusId: string) {
   const resourceStore = useResourceStore();
-  const { t } = useI18n();
   const { th } = useLocale();
   const mink = useMinkBackend();
-  const { alert } = useMessenger();
 
   const corpus = computed(() => resourceStore.corpora[corpusId]);
   const config = computed(() => corpus.value?.config);
-  const corpusName = computed(() => th(config.value?.name));
+  const configOptions = computedAsync(getParsedConfig);
+  const corpusName = computed(() => th(configOptions.value?.name));
 
   async function loadConfig() {
     const config = await mink
       .loadConfig(corpusId)
-      .then(async (yaml: string) => {
-        try {
-          return await parseConfig(yaml);
-        } catch (e) {
-          console.error(`Parsing config failed: ${e}`);
-          alert(t("corpus.config.parse.error"));
-          return undefined;
-        }
-      })
       // 404 means no config which is fine, rethrow other errors.
       .catch((error) => {
         if (error.response?.status == 404) return undefined;
@@ -42,16 +30,34 @@ export default function useConfig(corpusId: string) {
     corpus.value.config = config;
   }
 
-  async function uploadConfig(config: ConfigOptions) {
-    // This may throw, either from makeConfig or saveConfig.
-    await mink.saveConfig(corpusId, await makeConfig(corpusId, config));
-    resourceStore.corpora[corpusId].config = config;
+  async function uploadConfig(configOptions: ConfigOptions) {
+    const configYaml = await makeConfig(corpusId, configOptions);
+    await uploadConfigRaw(configYaml);
+  }
+
+  async function uploadConfigRaw(configYaml: string) {
+    await mink.saveConfig(corpusId, configYaml);
+    // Backend may modify uploaded config. Store our version immediately, but also fetch the real one unawaited.
+    resourceStore.corpora[corpusId].config = configYaml;
+    loadConfig();
+  }
+
+  async function getParsedConfig() {
+    if (!config.value) return undefined;
+    try {
+      const parsed = await parseConfig(config.value);
+      return parsed;
+    } catch (error) {
+      console.error(`Error parsing config for "${corpusId}":`, error);
+    }
   }
 
   return {
     config,
+    configOptions,
     corpusName,
     loadConfig,
     uploadConfig,
+    uploadConfigRaw,
   };
 }