+++ src/corpus/createCorpus.composable.ts
@@ -62,10 +62,10 @@ export default function useCreateCorpus() {
   }
 
   // Like the `uploadConfig` in `config.composable.ts` but takes `corpusId` as argument.
-  async function uploadConfig(config: ConfigOptions, corpusId: string) {
-    // This may throw, either from makeConfig or saveConfig.
-    await mink.saveConfig(corpusId, await makeConfig(corpusId, config));
-    resourceStore.corpora[corpusId].config = config;
+  async function uploadConfig(configOptions: ConfigOptions, corpusId: string) {
+    const configYaml = await makeConfig(corpusId, configOptions);
+    await mink.saveConfig(corpusId, configYaml);
+    resourceStore.corpora[corpusId].config = configYaml;
   }
 
   // Like the `uploadSources` in `sources.composable.ts` but takes `corpusId` as argument.