+++ src/router/corpus.routes.ts
@@ -4,6 +4,8 @@ const CorpusView = () => import("@/corpus/CorpusView.vue");
 const CorpusOverview = () => import("@/corpus/CorpusOverview.vue");
 const CorpusConfiguration = () =>
   import("@/corpus/config/CorpusConfiguration.vue");
+const CorpusConfigCustom = () =>
+  import("@/corpus/config/CorpusConfigCustom.vue");
 const CorpusResult = () => import("@/corpus/exports/CorpusResult.vue");
 const CorpusDelete = () => import("@/corpus/CorpusDelete.vue");
 const SourceView = () => import("@/corpus/sources/SourceView.vue");
@@ -35,6 +37,11 @@ const corpusRoutes: RouteRecordRaw[] = [
         path: "config",
         component: CorpusConfiguration,
         meta: { title: "configuration" },
+      },
+      {
+        path: "config/custom",
+        component: CorpusConfigCustom,
+        meta: { title: "config.custom" },
       },
       {
         path: "sources/:filename",